Introduce UseStmt.{h,cpp} based by extracting code from stmt.cpp and scopeResolve.cpp
