+++ compiler/AST/stmt.cpp
@@ -55,230 +55,6 @@ bool Stmt::isStmt() const {
   return true;
 }
 
-
-/******************************** | *********************************
-*                                                                   *
-*                                                                   *
-********************************* | ********************************/
-
-UseStmt::UseStmt(BaseAST* source):
-  Stmt(E_UseStmt),
-  src(NULL),
-  named(),
-  renamed(),
-  except(false),
-  relatedNames()
-{
-  if (Symbol* b = toSymbol(source)) {
-    src = new SymExpr(b);
-  } else if (Expr* b = toExpr(source)) {
-    src = b;
-  } else {
-    INT_FATAL(this, "Bad mod in UseStmt constructor");
-  }
-
-  gUseStmts.add(this);
-}
-
-//
-UseStmt::UseStmt(BaseAST*                            source,
-                 std::vector<const char*>*           args,
-                 bool                                exclude,
-                 std::map<const char*, const char*>* renames) :
-  Stmt(E_UseStmt),
-  src(NULL),
-  named(),
-  renamed(),
-  except(exclude),
-  relatedNames()
-{
-  if (Symbol* b = toSymbol(source)) {
-    src = new SymExpr(b);
-  } else if (Expr* b = toExpr(source)) {
-    src = b;
-  } else {
-    INT_FATAL(this, "Bad mod in UseStmt constructor");
-  }
-
-  if (args->size() > 0) {
-    // Symbols to search when going through this module's scope from an outside
-    // scope
-    for_vector(const char, str, *args) {
-      named.push_back(str);
-    }
-  }
-
-  if (renames->size() > 0) {
-    // The new names of symbols in the module being used, to avoid conflicts
-    // for instance.
-    for (std::map<const char*, const char*>::iterator it = renames->begin();
-         it != renames->end(); ++it) {
-      renamed[it->first] = it->second;
-    }
-  }
-
-  gUseStmts.add(this);
-}
-
-
-UseStmt* UseStmt::copyInner(SymbolMap* map) {
-  UseStmt *_this = 0;
-  if (named.size() > 0) { // MPF: should this have || renamed.size() > 0?
-    _this = new UseStmt(COPY_INT(src), &named, except, &renamed);
-  } else {
-    _this = new UseStmt(COPY_INT(src));
-  }
-  for_vector(const char, sym, relatedNames) {
-    _this->relatedNames.push_back(sym);
-  }
-  return _this;
-}
-
-void UseStmt::verify() {
-  Expr::verify();
-  if (astTag != E_UseStmt) {
-    INT_FATAL(this, "Bad NamedExpr::astTag");
-  }
-  if (src == NULL) {
-    INT_FATAL(this, "Bad UseStmt::src");
-  }
-  if (relatedNames.size() != 0 && named.size() == 0 && renamed.size() == 0) {
-    INT_FATAL(this, "Have names to avoid, but nothing was listed in the use to begin with");
-  }
-  verifyNotOnList(src);
-}
-
-void UseStmt::replaceChild(Expr* old_ast, Expr* new_ast) {
-  if (old_ast == src) {
-    src = new_ast;
-  } else {
-    INT_FATAL(this, "Unexpected case in UseStmt::replaceChild");
-  }
-}
-
-Expr* UseStmt::getFirstExpr() {
-  return this;
-}
-
-Expr* UseStmt::getFirstChild() {
-  return NULL;
-}
-
-void UseStmt::accept(AstVisitor* visitor) {
-  visitor->visitUseStmt(this);
-}
-
-void UseStmt::writeListPredicate(FILE* mFP) {
-  if (hasOnlyList()) {
-    fprintf(mFP, " 'only' ");
-  } else if (hasExceptList()) {
-    fprintf(mFP, " 'except' ");
-  }
-}
-
-bool UseStmt::hasOnlyList() {
-  return !isPlainUse() && !except;
-}
-
-bool UseStmt::hasExceptList() {
-  return !isPlainUse() && except;
-}
-
-bool UseStmt::isPlainUse() {
-  // This is an unmodified use statement if no 'only' or 'except' list was
-  // provided.
-  return named.size() == 0 && renamed.size() == 0;
-}
-
-// Return whether the use permits us to search for a symbol with the given
-// name.  Returns true ("should skip") if the name is related to our 'except'
-// list, or not present when we've been given an 'only' list.
-bool UseStmt::skipSymbolSearch(const char* name) {
-  if (isPlainUse()) {
-    // The use is unmodified by an 'except' or 'only' list, so it is safe to
-    // search for this name
-    return false;
-  }
-
-  if (except) {
-    // If the name is present in our 'except' list, or is a (type) constructor
-    // on a type that is in that list, or is a method or field on a type that
-    // is in that list, then we shouldn't look in this use for that name.
-    // Otherwise, it is safe to look.
-    if (matchedNameOrConstructor(name)) {
-      return true;
-    } else if (inRelatedNames(name)) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    // If the name is present in our 'only' list, or is a (type) constructor
-    // on a type that is in that list, or is a method or field on a type that
-    // is in that list, then we should look in this use for that name.
-    // Otherwise, we shouldn't look for that name here.
-    if (matchedNameOrConstructor(name)) {
-      return false;
-    } else if (inRelatedNames(name)) {
-      return false;
-    } else {
-      return true;
-    }
-  }
-}
-
-bool UseStmt::matchedNameOrConstructor(const char* name) {
-  for_vector(const char, toCheck, named) {
-    if (!strcmp(name, toCheck)) {
-      return true;
-    }
-  }
-  for(std::map<const char*, const char*>::iterator it = renamed.begin();
-      it != renamed.end(); ++it) {
-    if (!strcmp(name, it->first)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// Returns true if the name was in the relatedNames field, false otherwise.
-bool UseStmt::inRelatedNames(const char* name) {
-  for_vector(const char, toCheck, relatedNames) {
-    if (!strcmp(name, toCheck))
-      return true;
-  }
-  return false;
-}
-
-bool UseStmt::isARename(const char* name) {
-  return renamed.count(name) == 1;
-}
-
-const char* UseStmt::getRename(const char* name) {
-  return renamed[name];
-}
-
-// Should only be called when the mod field has been resolved
-BaseAST* UseStmt::getSearchScope() {
-  if (SymExpr* se = toSymExpr(src)) {
-    if (ModuleSymbol* module = toModuleSymbol(se->symbol())) {
-      return module->block;
-    } else if (TypeSymbol* enumTypeSym = toTypeSymbol(se->symbol())) {
-      // Assumes we have correctly verified that the type was an enum.
-      return enumTypeSym;
-    } else {
-      // Internal failure because resolving the mod field should have raised
-      // an error if it wasn't an enum or module
-      INT_FATAL(this, "Use invalid, not applied to module or enum");
-    }
-  } else {
-    INT_FATAL(this, "getSearchScope called before this use was processed");
-  }
-  return NULL;
-}
-
-
 /************************************* | **************************************
 *                                                                             *
 *                                                                             *
