+++ compiler/include/UseStmt.h
@@ -1,0 +1,103 @@
+/*
+ * Copyright 2004-2017 Cray Inc.
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef _USE_STMT_H_
+#define _USE_STMT_H_
+
+#include "stmt.h"
+
+class UseStmt : public Stmt {
+public:
+                  UseStmt(BaseAST* source);
+
+                  UseStmt(BaseAST*                            source,
+                          std::vector<const char*>*           args,
+                          bool                                exclude,
+                          std::map<const char*, const char*>* renames);
+
+  DECLARE_COPY(UseStmt);
+
+  virtual Expr*   getFirstExpr();
+
+  virtual Expr*   getFirstChild();
+
+  virtual void    replaceChild(Expr* oldAst, Expr* newAst);
+
+  virtual void    accept(AstVisitor* visitor);
+
+  virtual void    verify();
+
+  virtual GenRet  codegen();
+
+  bool            isPlainUse()                                           const;
+
+  bool            hasOnlyList()                                          const;
+
+  bool            hasExceptList()                                        const;
+
+  bool            isARename(const char* name)                            const;
+
+  const char*     getRename(const char* name);
+
+  bool            isValid()                                              const;
+
+  void            scopeResolve();
+
+  void            validateList();
+
+  UseStmt*        applyOuterUse(UseStmt* outer);
+
+  bool            skipSymbolSearch(const char* name)                     const;
+
+  bool            providesNewSymbols(UseStmt* other)                     const;
+
+  BaseAST*        getSearchScope()                                       const;
+
+  void            writeListPredicate(FILE* mFP)                          const;
+
+private:
+  bool            isValid(Expr* expr)                                    const;
+
+  Symbol*         getUsedSymbol(Expr* expr);
+
+  bool            isValidUsedSymbol(Symbol* symbol)                      const;
+
+  void            createRelatedNames(Symbol* maybeType);
+
+  bool            matchedNameOrConstructor(const char* name)             const;
+
+  bool            inRelatedNames(const char* name)                       const;
+
+  void            noRepeats()                                            const;
+
+  void            printUseError()                                        const;
+
+  void            printUseError(Symbol* sym)                             const;
+
+public:
+  Expr*                              src;
+  std::vector<const char*>           named;
+  std::map<const char*, const char*> renamed;
+
+private:
+  bool                               except;
+  std::vector<const char*>           relatedNames;
+};
+
+#endif
